refactor(reader): move every format reader into one internal package
