+++ internal/reader/json.go
@@ -1,0 +1,202 @@
+package reader
+
+import (
+	"bufio"
+	"bytes"
+	"encoding/json"
+	"errors"
+	"io"
+	"strings"
+
+	"github.com/nao1215/filesql/internal/infer"
+)
+
+// JSONDataColumn is the one column a JSON or JSONL table has: the raw JSON of
+// each element, held as text.
+//
+// The document is stored as it was written and read back with SQLite's
+// json_extract(), which is what makes an arbitrarily nested document queryable
+// without a schema:
+//
+//	SELECT json_extract(data, '$.address.city') FROM my_json_table;
+const JSONDataColumn = "data"
+
+// jsonHeader is the header every JSON and JSONL table has.
+func jsonHeader() ([]string, []infer.Type) {
+	return []string{JSONDataColumn}, []infer.Type{infer.Text}
+}
+
+// readJSON reads one JSON document into the "data" column. An array is streamed
+// element by element; any other value is one row.
+func readJSON(src io.Reader, opts Options, emit Emit) (Result, error) {
+	buffered := bufio.NewReader(src)
+	header, types := jsonHeader()
+	result := Result{Header: header, Types: types}
+
+	isArray, empty, err := peekJSONIsArray(buffered)
+	if err != nil {
+		return Result{}, err
+	}
+	if empty {
+		// A document holding nothing is a table with no rows. Only the caller
+		// knows whether that is a failure, so it is reported rather than refused.
+		result.EmptyInput = true
+		return result, emit(&Chunk{Header: header, Types: types})
+	}
+
+	if isArray {
+		decoder := json.NewDecoder(buffered)
+		if _, err := decoder.Token(); err != nil {
+			return Result{}, invalidError(err, "failed to parse JSON array")
+		}
+		rows, err := readJSONArray(decoder, opts, emit)
+		if err != nil {
+			return Result{}, err
+		}
+		result.Rows = rows
+		result.Total = rows
+		return result, nil
+	}
+
+	content, err := io.ReadAll(buffered)
+	if err != nil {
+		return Result{}, parseError(err, "failed to read JSON")
+	}
+	var value json.RawMessage
+	if err := json.Unmarshal(bytes.TrimSpace(content), &value); err != nil {
+		return Result{}, invalidError(err, "failed to parse JSON")
+	}
+	result.Rows = 1
+	result.Total = 1
+	return result, emit(&Chunk{Header: header, Records: [][]string{{string(value)}}, Types: types})
+}
+
+// peekJSONIsArray reports whether the document opens with '[', and whether it
+// holds no value at all. Leading whitespace is skipped.
+//
+// What the document opens with decides the shape, rather than whether it
+// unmarshals into a slice: "null" unmarshals into one as the empty slice, and
+// answering "empty JSON array" for a document holding no array named something
+// the caller had not written.
+func peekJSONIsArray(buffered *bufio.Reader) (isArray, empty bool, err error) {
+	for {
+		b, err := buffered.ReadByte()
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				return false, true, nil
+			}
+			return false, false, parseError(err, "failed to read JSON")
+		}
+		if b == ' ' || b == '\t' || b == '\n' || b == '\r' {
+			continue
+		}
+		// Put the byte back so the decoder or ReadAll can consume it.
+		if err := buffered.UnreadByte(); err != nil {
+			return false, false, parseError(err, "failed to read JSON")
+		}
+		return b == '[', false, nil
+	}
+}
+
+// readJSONArray streams the elements of an array whose opening bracket the
+// decoder has already consumed, and returns how many rows it emitted.
+func readJSONArray(decoder *json.Decoder, opts Options, emit Emit) (int, error) {
+	header, types := jsonHeader()
+	chunkSize := chunkSizeOf(opts)
+
+	var chunk [][]string
+	rows := 0
+	emitted := false
+	for decoder.More() {
+		var element json.RawMessage
+		if err := decoder.Decode(&element); err != nil {
+			return 0, parseError(err, "failed to decode JSON array element")
+		}
+		chunk = append(chunk, []string{string(element)})
+		if len(chunk) >= chunkSize {
+			rows += len(chunk)
+			if err := emit(&Chunk{Header: header, Records: chunk, Types: types}); err != nil {
+				return 0, err
+			}
+			chunk = nil
+			emitted = true
+		}
+	}
+
+	// Consume the closing bracket, then refuse anything after it ("[1] garbage").
+	if _, err := decoder.Token(); err != nil {
+		return 0, parseError(err, "failed to read JSON array end")
+	}
+	if decoder.More() {
+		return 0, invalidError(nil, "unexpected data after JSON array")
+	}
+
+	// An empty array still makes its table.
+	if len(chunk) > 0 || !emitted {
+		rows += len(chunk)
+		if err := emit(&Chunk{Header: header, Records: chunk, Types: types}); err != nil {
+			return 0, err
+		}
+	}
+	return rows, nil
+}
+
+// readJSONL reads one JSON value per line into the "data" column. A blank line
+// is not a value and is skipped.
+func readJSONL(src io.Reader, opts Options, emit Emit) (Result, error) {
+	buffered := bufio.NewReader(src)
+	header, types := jsonHeader()
+	result := Result{Header: header, Types: types}
+	chunkSize := chunkSizeOf(opts)
+
+	var chunk [][]string
+	emitted := false
+	lineNum := 0
+	for {
+		raw, err := buffered.ReadBytes('\n')
+		// Whatever arrived is processed before the error is checked: ReadBytes
+		// returns the last line together with io.EOF.
+		lineNum++
+		line := strings.TrimSpace(string(raw))
+		if line != "" {
+			if !json.Valid([]byte(line)) {
+				return Result{}, invalidError(nil, "invalid JSON on line %d: %s", lineNum, truncateLine(line, 100))
+			}
+			chunk = append(chunk, []string{line})
+			if len(chunk) >= chunkSize {
+				result.Rows += len(chunk)
+				if err := emit(&Chunk{Header: header, Records: chunk, Types: types}); err != nil {
+					return Result{}, err
+				}
+				chunk = nil
+				emitted = true
+			}
+		}
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				break
+			}
+			return Result{}, parseError(err, "failed to read JSONL")
+		}
+	}
+
+	// An input with no lines is a table with no rows. Only the caller knows
+	// whether that is a failure.
+	if len(chunk) > 0 || !emitted {
+		result.Rows += len(chunk)
+		if err := emit(&Chunk{Header: header, Records: chunk, Types: types}); err != nil {
+			return Result{}, err
+		}
+	}
+	result.Total = result.Rows
+	result.EmptyInput = result.Rows == 0
+	return result, nil
+}
+
+// truncateLine shortens a line for an error message.
+func truncateLine(s string, maxLen int) string {
+	if len(s) <= maxLen {
+		return s
+	}
+	return s[:maxLen] + "..."
+}
