+++ internal/reader/line_ending_test.go
@@ -1,4 +1,4 @@
-package parser
+package reader
 
 import (
 	"errors"
