+++ parser/json_test.go
@@ -461,31 +461,3 @@ func TestParse_JSONL_LargeLines(t *testing.T) {
 		assert.True(t, json.Valid([]byte(result.Records[1][0])))
 	})
 }
-
-func TestTruncateLine(t *testing.T) {
-	t.Parallel()
-
-	t.Run("short string unchanged", func(t *testing.T) {
-		t.Parallel()
-
-		result := truncateLine("hello", 10)
-
-		assert.Equal(t, "hello", result)
-	})
-
-	t.Run("long string truncated with ellipsis", func(t *testing.T) {
-		t.Parallel()
-
-		result := truncateLine("this is a very long string", 10)
-
-		assert.Equal(t, "this is a ...", result)
-	})
-
-	t.Run("exact length unchanged", func(t *testing.T) {
-		t.Parallel()
-
-		result := truncateLine("12345", 5)
-
-		assert.Equal(t, "12345", result)
-	})
-}
